+++ demo/site/server.js
@@ -8,6 +8,7 @@ const dev = process.env.NODE_ENV !== "production";
 const hostname = "localhost";
 const port = process.env.APP_PORT ?? 3000;
 const cdnEnabled = process.env.CDN_ENABLED === "true";
+const disableCdnOriginHeaderCheck = process.env.DISABLE_CDN_ORIGIN_HEADER_CHECK === "true";
 const cdnOriginHeader = process.env.CDN_ORIGIN_HEADER;
 
 // when using middleware `hostname` and `port` must be provided below
@@ -22,7 +23,7 @@ app.prepare()
                 // This tells it to parse the query portion of the URL.
                 const parsedUrl = parse(req.url, true);
 
-                if (cdnEnabled) {
+                if (cdnEnabled && !disableCdnOriginHeaderCheck) {
                     const incomingCdnOriginHeader = req.headers["x-cdn-origin-check"];
                     if (cdnOriginHeader !== incomingCdnOriginHeader) {
                         res.statusCode = 403;