+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## NEXT
+
+_tbd_
+
+### @comet/cms-api
+
+#### Changes
+
+-   Fix `createRedirect` mutation when no scoping is used
+
 ## 3.2.2
 
 _Jan 25, 2023_